# apps/api/src/swagger.yaml
openapi: 3.0.3
info:
  title: AgriTrack Central API
  description: |
    AgriTrack Central Backend API for Smart CRM Machinery Monitoring.
    
    ## Features
    - **Machines**: Real-time machine monitoring and management
    - **Bookings**: Farmer booking system
    - **Analytics**: Dashboard statistics
    - **Notifications**: SMS, Push, and Email alerts
    - **Geofencing**: Location-based boundaries and breach detection
    - **Fuel Tracking**: Real-time fuel monitoring and theft detection
    - **Maintenance**: Scheduled maintenance management
    
    ## WebSocket Events
    Connect to `ws://localhost:3001` for real-time updates:
    - `device_update`: Real-time machine state
    - `booking_update`: Booking status changes
    - `initial_state`: All machines on connect
    
  version: 2.0.0
  contact:
    name: AgriTrack Team
    url: https://github.com/KD-3030/AgriTrack
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/v1/machines/{id}:
    get:
      tags:
        - Machines
      summary: Get machine details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Machine details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
...
  /api/v1/geofences:
    get:
      tags:
        - Geofences
      summary: List all geofences
      responses:
        '200':
          description: List of geofences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Geofence'
    post:
      tags:
        - Geofences
      summary: Create a geofence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceRequest'
      responses:
        '201':
          description: Geofence created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
  /api/v1/geofences/{id}:
    get:
      tags:
        - Geofences
      summary: Get geofence with machines
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Geofence details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
    put:
      tags:
        - Geofences
      summary: Update geofence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceRequest'
      responses:
        '200':
          description: Geofence updated
    delete:
      tags:
        - Geofences
      summary: Delete geofence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Geofence deleted
  /api/v1/geofences/{id}/breaches:
    get:
      tags:
        - Geofences
      summary: Get breach history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Breach history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeofenceBreach'
  /api/v1/fuel/{machineId}:
    get:
      tags:
        - Fuel
      summary: Get fuel stats for machine
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fuel statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuelStats'
  /api/v1/fuel/{machineId}/history:
    get:
      tags:
        - Fuel
      summary: Get fuel history
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Fuel history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FuelReading'
  /api/v1/fuel/{machineId}/refueling:
    get:
      tags:
        - Fuel
      summary: Get refueling events
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Refueling events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RefuelingEvent'
    post:
      tags:
        - Fuel
      summary: Record refueling event
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                liters:
                  type: number
                cost:
                  type: number
                odometer:
                  type: number
      responses:
        '201':
          description: Refueling recorded
...
  /api/v1/maintenance/machine/{id}:
    get:
      tags:
        - Maintenance
      summary: Get machine maintenance schedules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Machine maintenance schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceSchedule'
...
components:
  schemas:
    Machine:
      type: object
      properties:
        id:
          type: string
        device_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [tractor, harvester, seeder, sprayer]
        status:
          type: string
          enum: [active, idle, maintenance, offline]
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
...
    Geofence:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [circle, polygon]
        center_lat:
          type: number
        center_lng:
          type: number
        radius_meters:
          type: number
        polygon_coords:
          type: object
        active:
          type: boolean
    GeofenceRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [circle, polygon]
        centerLat:
          type: number
        centerLng:
          type: number
        radiusMeters:
          type: number
        polygonCoords:
          type: object
    GeofenceBreach:
      type: object
      properties:
        id:
          type: string
        geofence_id:
          type: string
        device_id:
          type: string
        event_type:
          type: string
          enum: [entry, exit]
        timestamp:
          type: string
          format: date-time
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
    FuelStats:
      type: object
      properties:
        currentLevel:
          type: number
        consumptionRate:
          type: number
        estimatedRange:
          type: number
        lastRefuel:
          type: string
          format: date-time
    FuelReading:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: number
        consumption:
          type: number
    RefuelingEvent:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        liters:
          type: number
        cost:
          type: number
        odometer:
          type: number
...
    MaintenanceSchedule:
      type: object
      properties:
        id:
          type: string
        machine_id:
          type: string
        type:
          type: string
        due_date:
          type: string
          format: date
        status:
          type: string
          enum: [scheduled, overdue, completed]
        notes:
          type: string
...